new mutation called from frontend + ui branches for cognito role
