+++ types/queries/connected_forms.sql
@@ -46,7 +46,7 @@ from relations
       morpheme_gloss.id = relations.left_gloss_id or morpheme_gloss.id = relations.right_gloss_id
   inner join word_segment on word_segment.gloss_id = morpheme_gloss.id
   inner join word on word.id = word_segment.word_id
-  left join media_slice on media_slice.id = coalesce(word.curated_audio_slice_id, word.audio_slice_id)
+  left join media_slice on media_slice.id = word.audio_slice_id
   left join media_resource on media_resource.id = media_slice.resource_id
   left join dailp_user contributor on contributor.id = media_resource.recorded_by
   left join dailp_user editor on editor.id = word.audio_edited_by
