+++ website/src/components/subtle-button.tsx
@@ -1,0 +1,5 @@
+import { withClass } from "src/style/utils"
+import { CleanButton, IconTextButton } from "./button"
+import { subtleButton } from "./subtle-button.css"
+
+export const SubtleButton = withClass(subtleButton, CleanButton)
